requirements.txt
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
jupyterlab>=4.0
Pillow>=10.0
tk>=0.1.0
